requirements.txt
numpy
pandas
scikit-learn
matplotlib
tensorflow>=2.11
gym
stable-baselines3
mpi4py
newsapi-python
wordcloud
nltk
h5py
# keras  # Use tf.keras from TensorFlow 2.x
pytorch-lightning>=2.0
shap>=0.41.0
lime>=0.2.0.1
PyYAML>=6.0
websocket-client>=1.6.0